+++ templates/index.html
@@ -38,7 +38,7 @@
     </div>
 
     <div id="chat-section" class="hidden">
-        <form hx-post="/send" hx-target="#response" hx-swap="innerHTML" class="flex mb-4">
+        <form hx-post="/message/send" hx-target="#response" hx-swap="innerHTML" class="flex mb-4">
             <label class="w-100">
                 <input type="text" name="message" placeholder="Type your message" required
                        class="border border-gray-300 rounded-lg p-3 flex-grow focus:outline-none focus:ring-2 focus:ring-green-500">
@@ -49,7 +49,7 @@
             </button>
         </form>
         <div id="response" class="mb-4"></div>
-        <button hx-get="/history" hx-target="#chat-history" hx-swap="innerHTML"
+        <button hx-get="/message/history" hx-target="#chat-history" hx-swap="innerHTML"
                 class="bg-blue-500 text-white rounded-lg px-4 py-2 mb-4 transition duration-300 ease-in-out hover:bg-blue-400 focus:outline-none">
             Load History
         </button>
